+++ jest/jest.setup.ts
@@ -66,8 +66,8 @@ jest.mock('libs/subcategories/mappings')
 jest.mock('libs/subcategories/useCategoryId')
 jest.mock('libs/subcategories/useSubcategory')
 
-/* See the corresponding mock in libs/firebase/__mocks__ */
-jest.mock('libs/firebaseOld/getFirebaseApp')
+/* See the corresponding mock in libs/firebase/app/__mocks__ */
+jest.mock('libs/firebase/app/index.web')
 
 /* See the corresponding mocks in libs/campaign/__mocks__ */
 jest.mock('libs/campaign')
