+++ src/features/identityCheck/components/FastEduconnectConnectionRequestModal.tsx
@@ -9,7 +9,7 @@ import { openUrl } from 'features/navigation/helpers'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { analytics } from 'libs/analytics'
 import { env } from 'libs/environment'
-import { useUbbleETAMessage } from 'libs/firebaseOld/firestore/ubbleETAMessage'
+import { useUbbleETAMessage } from 'libs/firebaseImpl/firestore/ubbleETAMessage'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ButtonQuaternaryBlack } from 'ui/components/buttons/ButtonQuaternaryBlack'
 import { ButtonTertiaryBlack } from 'ui/components/buttons/ButtonTertiaryBlack'
