+++ src/libs/firebase/analytics/index.web.ts
@@ -1,9 +1,10 @@
-import { getFirebaseApp } from 'libs/firebase/app/getFirebaseApp'
+import { getFirebaseApp } from 'libs/firebase/app'
 
 const firebaseAnalytics = getFirebaseApp().analytics()
 
 export default () => ({
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  /* eslint-disable @typescript-eslint/no-unused-vars */
+  // @ts-ignore Not existing on the web
   setDefaultEventParameters(params: Record<string, string> | undefined) {
     // mock as not present on the web
   },
