+++ src/libs/firebase/app/__mocks__/index.web.ts
@@ -3,7 +3,9 @@ export const firebaseAnalytics = {
   setAnalyticsCollectionEnabled: jest.fn(),
   setUserId: jest.fn(),
 }
+
 const docSnapshopt = { get: jest.fn() }
+
 export const firebaseFirestore = {
   collection: jest.fn().mockReturnValue({
     doc: jest.fn().mockReturnValue({
