+++ src/libs/firebase/firestore/index.web.ts
@@ -1,1 +1,3 @@
-export default {}
+import { getFirebaseApp } from 'libs/firebase/app'
+
+export default () => getFirebaseApp().firestore()
